requirements.txt
streamlit
cryptography
gspread
google-auth
pandas
plotly
requirements.txt
